msm: kgsl: Check if underlying vma is cached
When importing a buffer via its user address, the user may not
specify (via passed in flags) that whether this mapping needs to be
io-coherent or not. This can cause the gpu mapping to be not marked
as io-coherent even though the underlying vma is cached. This can lead
to coherency issues. Hence, explicitly check if the underlying vma
is cached. If so, then mark the GPU side mapping as io-coherent.

Change-Id: Icd849902a403b02d3eec438d992d00c0a7fe23ee
Signed-off-by: Harshdeep Dhatt <quic_hdhatt@quicinc.com>