kgsl: Fix AB voting in gen8_3_0
When GMU-based AB voting is enabled, the vote calculation relies
on the number of DDR channels. Currently, this value is hard-coded
for each generation. However, it can vary from target to target
within the same generation. Thus, introduce a new member in gpu core
structure to specify the number of ddr channels for each target. If
GMU-based AB voting is supported but the number of DDR channels is
not specified, print a warning and fallback to interconnect path for
AB voting. This helps to fix the AB voting in gen8_3_0.

Change-Id: I83cad77d6acd246e9c45a40e20dcf186d38c9a2a
Signed-off-by: Kamal Agrawal <quic_kamaagra@quicinc.com>