asoc: update curr_micbias2 val with mb->micb2_mv
Curr_micbias2 val should be same with mb->micba2_mv.
for some customers, their micbias is not 1800mv,
so we should set the curr_micbias as mb->micb2_mv which
was the actual micbias val the customer use.
remove the dev_up check from micbias, since with this check,
Headset may be detected as a incorrect type.

Change-Id: Ia5a1fd434a09617a57096531c69956479c69f079
Signed-off-by: Yuhui Zhao <quic_yuhuzhao@quicinc.com>