msm: adsprpc: Avoid taking reference for group_info
Currently, the get_current_groups API accesses group info, which
increases the usage refcount. If the IOCTL using the
get_current_groups API is called many times, the usage counter
overflows. To avoid this, access group info without taking a
reference. A reference is not required as group info is not
released during the IOCTL call.

Change-Id: Iec31c90f9fd1a837fb697d5d7d1baba7d285374d
Signed-off-by: ANANDU KRISHNAN E <quic_anane@quicinc.com>